GH-13 Add verbosity to error message
This commit adds the prettyNames of the conflicting packages to the exception message. Now, instead of saying "Two packages cannot share the same directory!" it will tell you "Two packages (<name> and <name>) cannot share the same directory!". The first package name will be the package that is attempting to use a directory already used by the second package.

Props @GaryJones

Fixes #13